Cleanup some compilation products after call
When calling sadve, executables specific to the symbol being evaluated
are created. Try to remove as many of them as possible to not leave too
much rubbish after each call. On the other hand we cannot remove
everything as it significantly slows down the script execution time
becuase of cmake lokking for compilers, etc.