Introduce implicit mode
This is first step to implement offsetof support. If one of struct,
enum, union is selected, it is treated as implicit mode, that will be
committed and real mode only if real one was not selected explicitly
by user. type and new option offset are two examples of such explicit
mode. So if neither type nor offset was selected, mode is the one
selected implicitly with struct, enum or union, that is in this case
sizeof. However, if e.g. offset is selected as addition to struct, final
code that will be computed is offsetof(struct structname, field).